优化查询染料清单

+++ internal/app/dyeing-system-application/infrastructure/entity/dye.go
@@ -5,8 +5,8 @@ import "time"
 type Dye struct {
 	ID          uint64    `gorm:"primary_key;auto_increment;comment:主键" json:"id"`
 	Name        string    `gorm:"size:20;not null;comment:染料名" json:"name"`
-	TotalAmount int64     `gorm:"default:0;comment:总量" json:"total_amount"`
-	LastAmount  int64     `gorm:"default:0;comment:上次总量" json:"last_amount"`
+	TotalAmount float32   `gorm:"default:0;comment:总量" json:"total_amount"`
+	LastAmount  float32   `gorm:"default:0;comment:上次总量" json:"last_amount"`
 	Phone       string    `gorm:"size:15;default:0;comment:染料公司手机号" json:"phone"`
 	Company     string    `gorm:"size:100;default: ;comment:染料公司" json:"company"`
 	Address     string    `gorm:"size:100;default: ;comment:染料公司地址" json:"address"`
